data_engineering/abstract_factory/requirements.txt
mysql-connector-python
psycopg2-binary

